+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.0 (TBD)
 
 - [BREAKING] Remove `AccountIdAnchor` from account ID generation process (#1391).
+- Allow NOOP transactions and state-updating transactions against the same account in the same block (#1393).
 
 ## 0.9.0 (2025-05-20)